refactor: news feed container
