+++ src/renderer/pages/home/news_article/NewsArticle.tsx
