+++ src/renderer/pages/home/news_article/stories/NewsArticle.stories.tsx
