+++ src/renderer/pages/home/news_article/stories/fakeNewsItem.ts
